[SELC-5549] fix: Sonar issues in test classes
